Feat:영화 데이터 가져오는 함수 생성
